make consistent i32 for entity_Id and scene_entity_id clean dead code rename Promise.gd to promise.gd

+++ rust/decentraland-godot-lib/src/scene_runner/components/gltf_container.rs
@@ -63,7 +63,7 @@ pub fn update_gltf_container(scene: &mut Scene, crdt_state: &mut SceneCrdtState)
                     new_gltf
                         .bind_mut()
                         .set_dcl_gltf_src(GodotString::from(new_value.src));
-                    new_gltf.bind_mut().set_dcl_scene_id(scene_id as i32);
+                    new_gltf.bind_mut().set_dcl_scene_id(scene_id);
                     new_gltf.bind_mut().set_dcl_entity_id(entity.as_i32());
                     new_gltf
                         .bind_mut()
